Add basic structs and types to use going forward
