+++ internal/quiz/quiz.go
@@ -1,0 +1,8 @@
+package quiz
+
+import (
+	"context"
+	"github.com/jackc/pgx/v5"
+)
+
+type Question struct {
